rename model in models for consistency
