+++ mindbug-backend/src/main/java/com/mindbug/controller/CardController.java
@@ -2,7 +2,8 @@ package com.mindbug.controller;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.mindbug.model.Card;
+import com.mindbug.models.Card;
+
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
